+++ CHANGELOG.md
@@ -18,6 +18,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat(syslogexporter): change TLS configuration options to opentelemetry configtls [#983]
 - chore: bump golang to 1.19 [#1011]
 - chore: update OT core to v0.72.0 [#1013]
+- chore: add new components [#1016]
+
+  - datadogprocessor
+  - servicegraphprocessor
+  - awscloudwatchreceiver
+  - azureeventhubreceiver
+  - datadogreceiver
+  - haproxyreceiver
+  - iisreceiver
+  - k8sobjectsreceiver
+  - otlpjsonfilereceiver
+  - purefareceiver
+  - purefbreceiver
+  - solacereceiver
+  - sshcheckreceiver
+  - forwardconnector
+  - countconnector
 
 [#969]: https://github.com/SumoLogic/sumologic-otel-collector/pull/969
 [#975]: https://github.com/SumoLogic/sumologic-otel-collector/pull/975
@@ -25,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#983]: https://github.com/SumoLogic/sumologic-otel-collector/pull/983
 [#1011]: https://github.com/SumoLogic/sumologic-otel-collector/pull/1011
 [#1013]: https://github.com/SumoLogic/sumologic-otel-collector/pull/1013
+[#1016]: https://github.com/SumoLogic/sumologic-otel-collector/pull/1016
 [unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.71.0-sumo-0...main
 
 ## [v0.71.0-sumo-0]