+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v0.72.0-sumo-0]
+
+### Released 2023-02-24
 
 ### Added
 
@@ -46,7 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#1013]: https://github.com/SumoLogic/sumologic-otel-collector/pull/1013
 [#1016]: https://github.com/SumoLogic/sumologic-otel-collector/pull/1016
 [#1015]: https://github.com/SumoLogic/sumologic-otel-collector/pull/1015
-[unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.71.0-sumo-0...main
+[v0.72.0-sumo-0]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.71.0-sumo-0...v0.72.0-sumo-0
 
 ## [v0.71.0-sumo-0]